Persistent music added.
Since the persistent music object is in the scene SplashScreens.unity, you'll have to add the UndyingMusic archetype to your level to hear music while testing your level (remember to remove it once you are done).  You'll have to set the music for your level in HORRIBLESCRIPT (in LevelSettings) if it has not been done already.  Not having a LevelSettings object (with HORRIBLESCRIPT) in your scene will now cause a crash.