Was working on Conversation Persistence.
Apparently the Conversation System is going to be changed, so I'm going to wait before continuing and seeing all of my work rendered useless.  Don't worry, all of my stuff is in isolated scenes/the alternate timeline.

Other than that, minor bugfixes were also done.  So minor that I'm not going to mention them so that it all is unnoticed by everyone.  :^)

Goal for next week:  Be less tired, so that my commit messages are more readable and less inane dribble.