Rail Camera added: see description for instructions
Delete the old main camera, and instead use the RailCamera prefab. It
has 3 settings: Rail, Speed, and Look Direction.
Rail is the rail the camera will stick to.
Speed is the Lerp Speed (0 is no movement, 100 is match the player)
Look direction is the rotation (in euler angles) of the camera.

Use the Rail prefab to insert a 2 node rail. The camera will stay on a
point between these two nodes. You can duplicate nodes, and it will form
a rail between any duplicates as well. If A-B-C, A goes to B and B goes
to C, but A does not go to C. If C-A-B, C goes to A, A goes to B, but C
does not go to B.

Rail Camera Regions can move to a different rail, give a new move rate,
and/or a new look direction. If any of these are left at the default
value, they will be ignored. Once the player leaves the region, these
will reset to the defaults given on the camera.